[VB-14] style: changed package structure
